Long values for EMC (#1709)
* All emc related numbers now stored in Longs. Fixed bug with Novas (now
you can break them without crash).

* Tests restored

* Remove changes not pertaining to long conversion

* Fix CustomConversionMapper tests

* Fix GraphMapperTest.testGenerateValuesFreeAlternativesWithNegativeIngredients
Double.MIN_VALUE is a positive number so use negative DOUBLE.MAX_VALUE

* Bring back UpdateWindowIntPKT for fields that are integers

* Fixed a couple compile issues I created in the last commit

* Fix overflow not converting tooltip to obfuscated

* Fix hidden arithmetic test and bring back a hidden form for doubles
There previously was a hidden form for fractions which would round

* Make the tests compare longs to longs instead of integers to longs

* Fix potential long overflow

* Fix GraphMapperTest.testOverflowWithIngredients
Properly handles when it overflows, double addition apparently just gives the max value instead of erroring about an overflow

* Actually use proper min value for DoubleArithmetic

Fixes GraphMapperTest.testGenerateValuesWaterBucketRecipe and GraphMapperTest.testGenerateValuesFreeAlternatives

* Convert long literal to capital L in tests for readability
Unfortunately the literal is needed as java cannot autopromote from int to a Long object

* Small formatting changes to make pull request more readable with less changes
Yes this adds some whitespace that is currently in the base but it makes it easier to read the changes because it doesn't show the dif for removing the spacing

* Better handling of long packet
Supports using updateProgressBarLong for ids other than emc but they should never end up calling it as no UpdateWindowLongPKTs are sent when it isn't dealing with emc
Also fixes a potential overflow in CondenserContainer.getProgressScaled by ensuring the longs are divided before multiplying a potentially max long by a constant for progression

* remainEMC is always a long so don't cast it to a double potentially losing accuracy

* Send the full value of EMC when SyncEmcPKT is used
This is done by sending the Most significant and least significant bits of the EMC long and then merging them back together during read

* Fix condenser progress bar to actually update

* Requested changes

* Switch to BigFraction and shade the jar with apache commons math3

* Fix a file being in the wrong package

* Cleanup gradle shading slightly to use correct jar name

* Exclude a lot of the unneeded commons-math3 stuff shrinks size to 1737KB

+++ src/main/java/moze_intel/projecte/api/proxy/IEMCProxy.java
@@ -66,7 +66,7 @@ public interface IEMCProxy
      * @param block The block we want to query
      * @return The block's EMC value, or 0 if there is none
      */
-    int getValue(@Nonnull Block block);
+    long getValue(@Nonnull Block block);
 
     /**
      * Queries the EMC value for the provided item
@@ -75,7 +75,7 @@ public interface IEMCProxy
      * @param item The item we want to query
      * @return The item's EMC value, or 0 if there is none
      */
-    int getValue(@Nonnull Item item);
+    long getValue(@Nonnull Item item);
 
     /**
      * Queries the EMC value for the provided stack
@@ -85,5 +85,5 @@ public interface IEMCProxy
      * @param stack The stack we want to query
      * @return The stack's EMC value, or 0 if there is none
      */
-    int getValue(@Nonnull ItemStack stack);
+    long getValue(@Nonnull ItemStack stack);
 }
