+++ src/main/java/moze_intel/projecte/emc/arithmetics/IValueArithmetic.java
@@ -9,8 +9,8 @@ public interface IValueArithmetic<T extends Comparable<T>> {
 	boolean isZero(T value);
 	T getZero();
 	T add(T a, T b);
-	T mul(int a, T b);
-	T div(T a, int b);
+	T mul(long a, T b);
+	T div(T a, long b);
 
 	/**
 	 * An item that has a "free" value contributes no additional cost to
