+++ src/main/java/moze_intel/projecte/emc/collector/DumpToFileCollector.java
@@ -10,14 +10,14 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Map;
 
-public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMappingCollector<NormalizedSimpleStack, Integer, A>
+public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMappingCollector<NormalizedSimpleStack, Long, A>
 {
 	public static String currentGroupName="default";
 	private final CustomConversionFile out = new CustomConversionFile();
-	private final IExtendedMappingCollector<NormalizedSimpleStack, Integer, A> inner;
+	private final IExtendedMappingCollector<NormalizedSimpleStack, Long, A> inner;
 	private final File file;
 
-	public DumpToFileCollector(File f, IExtendedMappingCollector<NormalizedSimpleStack, Integer, A> inner)
+	public DumpToFileCollector(File f, IExtendedMappingCollector<NormalizedSimpleStack, Long, A> inner)
 	{
 		super(inner.getArithmetic());
 		this.file = f;
@@ -43,7 +43,7 @@ public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMap
 	}
 
 	@Override
-	public void setValueBefore(NormalizedSimpleStack something, Integer value)
+	public void setValueBefore(NormalizedSimpleStack something, Long value)
 	{
 		inner.setValueBefore(something, value);
 		if (something == null) return;
@@ -51,7 +51,7 @@ public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMap
 	}
 
 	@Override
-	public void setValueAfter(NormalizedSimpleStack something, Integer value)
+	public void setValueAfter(NormalizedSimpleStack something, Long value)
 	{
 		inner.setValueAfter(something, value);
 		if (something == null) return;
