+++ src/main/java/moze_intel/projecte/emc/mappers/APICustomEMCMapper.java
@@ -20,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Integer> {
+public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Long> {
 	public static final APICustomEMCMapper instance = new APICustomEMCMapper();
 	private static final int PRIORITY_MIN_VALUE = 0;
 	private static final int PRIORITY_MAX_VALUE = 512;
@@ -28,15 +28,15 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 	private APICustomEMCMapper() {}
 
 	//Need a special Map for Items and Blocks because the ItemID-mapping might change, so we need to store modid:unlocalizedName instead of the NormalizedSimpleStack which only holds itemid and metadata
-	private final Map<String, Map<NormalizedSimpleStack, Integer>> customEMCforMod = new HashMap<>();
-	private final Map<String, Map<NormalizedSimpleStack, Integer>> customNonItemEMCforMod = new HashMap<>();
+	private final Map<String, Map<NormalizedSimpleStack, Long>> customEMCforMod = new HashMap<>();
+	private final Map<String, Map<NormalizedSimpleStack, Long>> customNonItemEMCforMod = new HashMap<>();
 
-	public void registerCustomEMC(ItemStack stack, int emcValue) {
+	public void registerCustomEMC(ItemStack stack, long emcValue) {
 		if (stack.isEmpty()) return;
 		if (emcValue < 0) emcValue = 0;
 		ModContainer activeMod = Loader.instance().activeModContainer();
 		String modId = activeMod == null ? null : activeMod.getModId();
-		Map<NormalizedSimpleStack, Integer> modMap;
+		Map<NormalizedSimpleStack, Long> modMap;
 		if (customEMCforMod.containsKey(modId)) {
 			modMap = customEMCforMod.get(modId);
 		} else {
@@ -46,13 +46,13 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 		modMap.put(NSSItem.create(stack), emcValue);
 	}
 
-	public void registerCustomEMC(Object o, int emcValue) {
+	public void registerCustomEMC(Object o, long emcValue) {
 		NormalizedSimpleStack stack = ConversionProxyImpl.instance.objectToNSS(o);
 		if (stack == null) return;
 		if (emcValue < 0) emcValue = 0;
 		ModContainer activeMod = Loader.instance().activeModContainer();
 		String modId = activeMod == null ? null : activeMod.getModId();
-		Map<NormalizedSimpleStack, Integer> modMap;
+		Map<NormalizedSimpleStack, Long> modMap;
 		if (customNonItemEMCforMod.containsKey(modId)) {
 			modMap = customNonItemEMCforMod.get(modId);
 		} else {
@@ -78,7 +78,7 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 	}
 
 	@Override
-	public void addMappings(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config) {
+	public void addMappings(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config) {
 		Map<String, Integer> priorityMap = new HashMap<>();
 		Set<String> modIdSet = new HashSet<>();
 		modIdSet.addAll(customEMCforMod.keySet());
@@ -118,7 +118,7 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 			String modIdOrUnknown = modId == null ? "unknown mod" : modId;
 			if (customEMCforMod.containsKey(modId))
 			{
-				for (Map.Entry<NormalizedSimpleStack, Integer> entry : customEMCforMod.get(modId).entrySet())
+				for (Map.Entry<NormalizedSimpleStack, Long> entry : customEMCforMod.get(modId).entrySet())
 				{
 					NormalizedSimpleStack normStack = entry.getKey();
 					if (isAllowedToSet(modId, normStack, entry.getValue(), config))
@@ -134,7 +134,7 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 			}
 			if (customNonItemEMCforMod.containsKey(modId))
 			{
-				for(Map.Entry<NormalizedSimpleStack, Integer> entry: customNonItemEMCforMod.get(modId).entrySet()) {
+				for(Map.Entry<NormalizedSimpleStack, Long> entry: customNonItemEMCforMod.get(modId).entrySet()) {
 					NormalizedSimpleStack normStack = entry.getKey();
 					if (isAllowedToSet(modId, normStack, entry.getValue(), config))
 					{
@@ -150,7 +150,7 @@ public class APICustomEMCMapper implements IEMCMapper<NormalizedSimpleStack, Int
 		}
 	}
 
-	private boolean isAllowedToSet(String modId, NormalizedSimpleStack stack, Integer value, Configuration config) {
+	private boolean isAllowedToSet(String modId, NormalizedSimpleStack stack, Long value, Configuration config) {
 		String itemName;
 		if (stack instanceof NSSItem)
 		{
