+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK3.java
@@ -36,7 +36,7 @@ public class GUICollectorMK3 extends GuiContainer
 	@Override
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
-		this.fontRenderer.drawString(Integer.toString(container.emc), 91, 32, 4210752);
+		this.fontRenderer.drawString(Long.toString(container.emc), 91, 32, 4210752);
 		
 		double kleinCharge = container.kleinEmc;
 		if (kleinCharge > 0)
