+++ src/main/java/moze_intel/projecte/gameObjs/items/GemEternalDensity.java
@@ -106,7 +106,7 @@ public class GemEternalDensity extends ItemPE implements IAlchBagItem, IAlchChes
 			}
 		}
 		
-		int value = EMCHelper.getEmcValue(target);
+		long value = EMCHelper.getEmcValue(target);
 
 		if (!EMCHelper.doesItemHaveEmc(target))
 		{
